refactor: Adicionando lógica necessaria para as questões 9 - 11
