+++ questao5a11/Gerente.java
@@ -1,0 +1,8 @@
+package questao5a11;
+
+public class Gerente extends Comissao {
+    public Gerente(){
+        super(1500);
+    }
+
+}
