+++ questao5a11/Supervisor.java
@@ -1,0 +1,10 @@
+package questao5a11;
+
+public class Supervisor extends Comissao {
+
+    public Supervisor(){
+        super(600);
+    }
+
+
+}
