+++ questao5a11/Vendedor.java
@@ -1,0 +1,7 @@
+package questao5a11;
+
+public class Vendedor extends Comissao{
+    public Vendedor(){
+        super(250);
+    }
+}
