--- questao5a11/questao8/Gerente.java
@@ -1,8 +1,0 @@
-package questao5a11.questao8;
-
-public class Gerente extends Comissao {
-    public Gerente(){
-        super(1500);
-    }
-
-}
