--- questao5a11/questao8/Supervisor.java
@@ -1,10 +1,0 @@
-package questao5a11.questao8;
-
-public class Supervisor extends Comissao {
-
-    public Supervisor(){
-        super(600);
-    }
-
-
-}
