--- questao5a11/questao8/Vendedor.java
@@ -1,7 +1,0 @@
-package questao5a11.questao8;
-
-public class Vendedor extends Comissao{
-    public Vendedor(){
-        super(250);
-    }
-}
